Refactor filter implementation
